teste

+++ includes/fdf_structs.h
@@ -6,7 +6,7 @@
 /*   By: ridalgo- <ridalgo-@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/25 19:51:55 by vkist-si          #+#    #+#             */
-/*   Updated: 2022/09/07 20:56:03 by ridalgo-         ###   ########.fr       */
+/*   Updated: 2022/09/07 22:03:48 by ridalgo-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -43,10 +43,10 @@ typedef struct s_win
 
 typedef struct s_dot
 {
-	float	x;
-	float	y;
-	float	z;
-	int		color;
+	int	x;
+	int	y;
+	int	z;
+	int	color;
 }	t_dot;
 
 typedef struct s_data
